Get tests to pass with the banana bomb wired in
Required that I treat the position of the current worm as possible
nothing, because a worm could be killed during the earlier steps of
the engine tick.  This was never a problem for the shoot move because
worms dying before you get to shooting only happens when both worms
moved and so fromJust in shoot would never have been called.
Nonetheless I decided to handle that case too.  Eventually I might
want to change everything to not use fromJust at all, but for now I'm
working with these assumptions.